+++ gateway/src/main/java/ru/practicum/shareit/booking/dto/BookItemRequestDto.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class BookItemRequestDto {
-	private long itemId;
+	private Integer itemId;
 	@FutureOrPresent
 	private LocalDateTime start;
 	@Future
